requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0.9
matplotlib>=3.8
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
opencv-python>=4.9
streamlit>=1.34
grad-cam>=1.5.0
pillow
opencv-python-headless